Bump deps to get things to work
.tool-versions: Had to bump to a newer erlang to build on macOS Big Sur; the latest
OTP 22 (22.3.4.16) led to weird compile errors, so I'm trying the latest 23. This
means I have to bump the elixir version too, to the latest version of the earliest
elixir built with OTP 23

hackney also had issues with OTP23 and needed to be bumped - TLS errors otherwise.

appsignal had compile errors and needed to be bumped to a later version where
Phoenix support was broken out into its own package.